requirements.txt
tensorflow==2.15.0 
Pillow==10.1.0
numpy==1.24.3
keras==2.15.0
flask==3.0.0
python-multipart==0.0.6
jinja2==3.1.2
openai
python-dotenv
streamlit